Add initial JS Game files and assets
Added the main HTML, CSS, and JavaScript files for a simple number guessing game, along with supporting images and assets. The game logic prompts the user to guess a secret number and provides feedback. Includes two sets of files: one in the root directory and another in the img/JS Game_files directory, as well as all required images for UI styling.